only add real files no directories

+++ usr/bin/archboot-tarball-helper.sh
@@ -54,7 +54,7 @@ mkdir -p ${TEMPDIR}/${BOOTDIRNAME}/
 # prepare syslinux bootloader
 install -m755 /usr/lib/syslinux/isolinux.bin ${TEMPDIR}/${BOOTDIRNAME}/isolinux.bin
 for i in /usr/lib/syslinux/*; do
-    install -m644 $i ${TEMPDIR}/${BOOTDIRNAME}/$(basename $i)
+    [ -f $i ] && install -m644 $i ${TEMPDIR}/${BOOTDIRNAME}/$(basename $i)
 done
 install -m644 /lib/modules/$(uname -r)/modules.pcimap ${TEMPDIR}/${BOOTDIRNAME}/modules.pcimap
 install -m644 /usr/share/hwdata/pci.ids ${TEMPDIR}/${BOOTDIRNAME}/pci.ids
